bug: Fix InterpreterManager's weird behaviour when reaching the last step
Save the StepResults for each visualisation in a separate list. This makes it easier to check if you're returning from a stepback and if the next step is the last, which allows you to block/unblock
the NextStep and NexSolution Buttons more easily when reaching the end of the visualisation.