qa: Add version number to gradle.build file
This will add the appropriate version number to the name of the .jar file when building the project.
Unfortunately, it isn't that easy to read the project version number from Java, therefore a version bump means a change to both the build.gradle file, as well as the MainWindow class.