qa: Wrap MainWindow constructor inside a blocking invokeAndWait() method
To prevent threading issues with Swing, the entire constructor is wrapped in an
invokeAndWait() method. This is arguably a bit ugly, however running JUnit tests
revealed that Swing does run the window initialization in a separate thread,
causing weird race conditions and producing NullPointer- and
ConcurrentModificationExceptions.

When starting the program as a user, you'll probably never run into these Swing
threading issues. However the unit tests do create a MainWindow instance and
immediately call getters an setters, relying on the MainWindow initialization to
be complete. The only way to garuantee that is to run everything inside a
blocking invokeAndWait() method.