qa: Use symbols for indicating whether a comparison succeeded
Because of a Java for Windows bug (TM), unicode symbols such as the check mark and the cross wouldn't be displayed correctly as shown in the Pflichtenheft. Until now, the visualisation would just show the words 'true' or 'false'. This commit re-introduces the symbols by using HTML entities such as '&#2717;' for the cross.